initial C code
The C code consists of a trivial library with three functions.
This commits adds the library (src/foo.c and src/foo.h), a program
that uses it, and a Makefile to build it.

The plan is to move two of them (foo_get and foo_set) to Rust, while
keeping the third (foo_print) in C.  Both C and Rust code must be
able to access the entire implementation, because the program will
remain in C, while the tests will be in Rust and will want to call
foo_print.

The conversion to Rust is pretty smooth, because there are no ownership
issues---the data structure is entirely self-contained.  There is only one
prerequisite for the API to be nicely usable from Rust code, which is
to use "const" properly in "foo_get", "foo_set" and "foo_print".