replace with Rust implementation
In this commit, the foo_get and foo_set functions switch to a Rust
implementation.  Even though bindgen still generates the prototypes
for them in "bindgen::foo", we do not re-export them from "foo".
Instead, "foo" implements the functions that C code will call.  They
are marked as "no_mangle" which is essentially the Rust equivalent
of C++'s extern "C".

Note that the implementation is unsafe; it has to be for two reasons.
First, because it dereferences a raw pointer ("let foo = &*foo" and
the similar assignment in foo_set) that could be pointing to garbage
or to freed data---and this is unsafe, so much that the code would not
compile at all if those dereferences were not "unsafe".  Second, because
we have no idea of what could be in the struct, for example it might
be uninitialized.  This is not something that the compiler tells us,
we have to figure it out!

Note that this is probably not a step that you would do explicitly,
in practice.  As long as you have tests for the C API written in either
Rust or C, you can switch directly to a Rust implementation as in the
next patch.