"rustify" implementation, with bindings provided as extern functions
Finally in this patch we complete the exercise and we are able to enjoy
the benefits of Rust in implementing the Foo data structure.

The data structure is still defined in a C header, and the
"bindgen::foo:Foo" struct that bindgen created for us is still what
gets re-exported as "foo::Foo".  However, it now grows Rust methods,
including a constructor "Foo::new" that was not even there in the original
C API, and the C API becomes a simple wrapper that converts the raw
pointers to C code and calls those methods.  In this commit we add a
second test, which calls those Rust methods instead of relying on the
C API.  Note that this test does not have to use any unsafe block,
and in fact none of the code added in this commit does.