add Rust bindings to the C library
In this commit, we start integrating the C and Rust code; in particular,
we keep the C implementation and add Rust tests for it.

The Rust library grows two modules, "foo" and "bindgen::foo".  The former
provides the public API, while the latter is the automatically generated
code from bindgen.  We do this already, even though for now the public API
is just forwarding the bindgen struct and entry points using "pub use".
Separating the automatically generated code into the bindgen module lets
us disable some errors that make no sense for bindgen code, so it's a
good idea.

We also need to link the C code into the Rust tests.  Doing this
requires adding a build.rs file.  This Rust source file produces two
directives for cargo, essentially corresponding to -L and -l on the
linker command line.