integrate Cargo into the Makefile
Here we add a Rust library, just like one for which "cargo new --lib"
creates the scaffolding, to the project's build system.
The Rust library creates another static library, and the Makefile
takes care of linking it into the main program.

For simplicity we will let Cargo run the build of the Rust parts, but
since the project is still first and foremost a C one, the overall build
will be driven by "make".  Therefore, after this commit you will be able
to do "make check" and see some Rust tests run.

At this point, the C code is already linking to the Rust library, but
the Rust library however is not able to use the C code; this will change
in the next commit.